Removes the Nanotrasen Asteroid Outpost (The one that's been disabled for like a year) (#3747)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Removes a disabled outpost that isn't intended to be brought back.
Already discussed with other maptainers and I'm tired of getting
questions about it.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This outpost hasn't been used ingame for a long time. It's been slated
for removal for quite a while now.

:cl:
del: Removed the unused NT_Asteroid outpost.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->