Gorlex Sledgehammer (#3718)
## About The Pull Request

Adds the Breaching Sledgehammer, a melee weapon utilized by the Gorlex
Marauder splinters both for combat and as a powerful demolition / mining
implement.

Sprites by @Imaginos16 


![image](https://github.com/user-attachments/assets/c9382aa2-8831-42ee-994f-4acbc6d8b22d)

Originally was going to make them do increased damage to structures but
its already 30 force it'll crack open most things with no issue

## Why It's Good For The Game

Gorlex uses mining tools like this sledgehammer in lore, its reflective
of their origins and is also cool as hell.

## Changelog

:cl:
add: Adds the Gorlex Marauder Breaching Sledgehammer, for both utility
and combat. Available in the black market.
add: Sledgehammers are mapped in on the Kali and Derecho-class
/:cl: