Surgery Cases (#3863)
adds surgery cases and a base case (useless right now)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new storage type called cases, and adds a surgical case to it for
your surgical needs. (no more dufflebag)
Also replaces all spawned surgical duffle bags
Buying surgical equipment now also gives you a health analyzer (included
with the case)

![image](https://github.com/user-attachments/assets/27ff5d5c-146b-4195-9a08-49e575244c68)

> Hey, the tools are even in the order most surgeries go, isn't that
cute.


![image](https://github.com/user-attachments/assets/2576228d-658e-4b2e-99d7-dc86ac2368ed)

> And yes, you can store your fancy tools in there, Cybersun

>_Finally, the PR will be on Draft until new sprites and or permission
can be obtained since the current ones I have absolutely no clue where
it came from and i'd rather not use it without crediting them. When the
sprites are here, they will be posted alongside this PR, and obviously
credited._

**GOOD NEWS**
I found the original spriter and obtained permission to use their
sprites.
# The sprite was done by the lovely **Azlan** 
And the original PR that added them to Baystation is mucked in with the
Torch Remap (as it was added with the aquila???)
However, this is the PR that brings us the Surgical Case as we know it
today.
https://github.com/Baystation12/Baystation12/pull/16637
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Doing surgery currently is immensely painful, as items are often hard to
grab resulting in patients quite literally bleeding out.
For example last round I played the doctor struggled for the hemostat so
much that my character almost bled out and died. Which, I shouldn't need
to explain being bad.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: added surgical cases which hold your tools
del: removed all naturally spawning surgical duffles
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->