Easy Energy Cells -- Integrated Retainment Latches (#3853)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

All energy weapons now come with a retainment clip for their energy
cells, if their cells are designed to be removable. After a cell is
unlatched, a user will be able to remove the cell like as they would a
magazine from a ballistic weapon. However, one must remember that these
retainment latches were designed into these weapons for a reason. Firing
one with a latch hanging loose will likely cause the cell to fall out
and/or disconnect due to jostling.

In order to latch or unlatch a retainment clip, one must alt-click on
their weapon.

If one is using a dual-mode weapon, such as an E-40, they will need to
switch to laser mode in order to latch and unlatch the retainment clip,
and remove the cell.

If one has attachments on their weapon, you'll need to remove the cell
on their weapon before attempting to remove an attachment. For the E-40,
one simply needs to swap to the ballistic mode in order to do that.

A line indicating the latch status was added to the examine text of
energy weapons.

![image](https://github.com/user-attachments/assets/5bd8d99b-c066-4664-a6e0-6584fe11667f)

This PR also slightly modifies the examine text of the E-40 to include
cell charge characteristics.

![image](https://github.com/user-attachments/assets/5bfadcd4-1ca7-4e10-9e56-c37f382e549e)

Small sample of it in operation:

https://github.com/user-attachments/assets/5ac44c60-fed4-4448-ac15-4f65cf3202b9

Full demo of it in action:
https://youtu.be/Be1c_0AFAds

A clip of the state of the latch sprites at this time:

https://github.com/user-attachments/assets/66d036c1-27df-486f-9b34-c82c22adbd06

Having cell removal be tied to alt-click felt smoother to me, compared
to the current way that cells are handled, with removal requiring a
screwdriver. The time it takes to remove and replace a cell is
comparable as before the change, but it no longer requires a screwdriver
to do.

:cl:
balance: replaced screwdriver cell removal with a cell-retainment clip
mechanism
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->