Shoulder Sling attachment (#3860)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Adds the shoulder sling attachment to cargo for 500 credits.

The shoulder sling can be attached any long arm gun with a weight class
of bulky or up. The sling allows for the gun to be carried in your suit
storage slot without an armor holster.

Adds support for attachment overlays to render under the attached gun.

Shoulder sling sprites by Moffball.


![image](https://github.com/user-attachments/assets/f6b9910f-b42b-4e38-87fa-a146061c05d5)

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
This gives players more options for fashion and accessorizing. You're no
longer stuck with wearing a boring armor vest to hold your bulky
weapons, and the other option of carrying your gun everywhere in your
hands makes you look like a maniac. Balance wise this should be fine,
since you can buy armor from the outpost anyways, which provides the
same benefit with it's holster while being objectively better because
well. It's armored.


## Changelog

:cl:
add: Shoulder slings to outpost cargo for 500 credits.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->